separate out super tw, make tips now or later a separate component as well, set up question library page tip
