+++ tutor/resources/styles/components/tours/value-prop.less
@@ -63,7 +63,6 @@
   }
   .btn {
     color: @tutor-neutral-lightest;
-    font-weight: 700;
     background-color: @tutor-tertiary-light;
     border-radius: 5px;
     padding: @spacing*0.25 @spacing;
@@ -110,7 +109,7 @@
   &.question-library {
     .column {
       background-size: auto 100px;
-      padding-top: 120px;
+      padding-top: 116px;
 
       &.exclude-question  { .tutor-background-image("question-library/exclude-question.svg"); }
       &.machine-learning  { .tutor-background-image("question-library/machine-learning.svg"); }
