+++ tutor/src/components/tours/custom/how-to-use-preview.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import {
   ValueProp,
   ColumnContent,
-  Column,
-  SuperTrainingWheel
+  Column
 } from './common';
+import SuperTrainingWheel from './super-training-wheel';
 
 export default class HowToUsePreview extends React.PureComponent {
