+++ tutor/src/components/tours/custom/index.js
@@ -1,11 +1,13 @@
-import HowToUseQL from './how-to-use-ql';
-import HowToUsePreview from './how-to-use-preview';
-import ValueProp from './value-prop';
-import AboutPageTips from './about-page-tips';
+import HowToUseQL               from './how-to-use-ql';
+import HowToUsePreview          from './how-to-use-preview';
+import ValueProp                from './value-prop';
+import DashboardPageTips        from './dashboard-page-tips';
+import QuestionLibraryPageTips  from './question-library-page-tips';
 
 export default {
   HowToUseQL,
   HowToUsePreview,
   ValueProp,
-  AboutPageTips
+  DashboardPageTips,
+  QuestionLibraryPageTips
 };
