+++ tutor/src/tours/question-library.json
@@ -10,10 +10,8 @@
     {
       "customComponent": "HowToUseQL"
     }, {
-      "body": "More about the Question Library in our “Page tips.”",
+      "customComponent": "QuestionLibraryPageTips",
       "anchor_id": "menu-option-page-tips",
-      "position": "bottom-right",
-      "is_fixed": true,
       "action": {
         "id": "open-support-menu"
       }
